+++ example/lib/pubspec.dart
@@ -1,0 +1,32 @@
+// This file is generated automatically, do not modify
+
+// ignore_for_file: public_member_api_docs, constant_identifier_names
+
+mixin Pubspec {
+  static final buildDate = DateTime.utc(2025, 6, 16, 18, 9, 19);
+
+  static const name = 'example';
+
+  static const description = 'An example of using dartantic_ai.';
+
+  static const publish_to = 'none';
+
+  static const environment = <dynamic, dynamic>{'sdk': '^3.7.2'};
+
+  static const dependencies = <dynamic, dynamic>{
+    'dartantic_ai': <dynamic, dynamic>{'path': '..'},
+    'dotprompt_dart': '^0.2.0',
+    'google_generative_ai': '^0.4.7',
+    'http': '^1.2.0',
+    'json_annotation': '^4.9.0',
+    'openai_dart': '^0.4.5',
+    'soti_schema': '^1.0.0',
+    'timezone': '^0.10.1',
+  };
+
+  static const dev_dependencies = <dynamic, dynamic>{
+    'all_lint_rules_community': '^0.0.43',
+    'build_runner': '^2.4.15',
+    'json_serializable': '^6.9.0',
+  };
+}