+++ lib/pubspec.dart
@@ -1,0 +1,50 @@
+// This file is generated automatically, do not modify
+
+// ignore_for_file: public_member_api_docs, constant_identifier_names
+
+mixin Pubspec {
+  static final buildDate = DateTime.utc(2025, 6, 16, 18, 9, 19);
+
+  static const name = 'dartantic_ai';
+
+  static const description =
+      'A Dart library for easily interacting with LLMs in a typed manner.';
+
+  static const versionFull = '0.5.0';
+
+  static const version = '0.5.0';
+
+  static const versionSmall = '0.5';
+
+  static const versionMajor = 0;
+
+  static const versionMinor = 5;
+
+  static const versionPatch = 0;
+
+  static const versionBuild = 0;
+
+  static const versionPreRelease = '';
+
+  static const versionIsPreRelease = false;
+
+  static const repository = 'https://github.com/csells/dartantic_ai';
+
+  static const environment = <dynamic, dynamic>{'sdk': '^3.7.2'};
+
+  static const dependencies = <dynamic, dynamic>{
+    'dotprompt_dart': '^0.2.0',
+    'google_generative_ai': '^0.4.7',
+    'json_schema': '^5.2.1',
+    'mcp_dart': '^0.5.1',
+    'openai_dart': '^0.4.5',
+    'uuid': '^4.3.3',
+  };
+
+  static const dev_dependencies = <dynamic, dynamic>{
+    'all_lint_rules_community': '^0.0.43',
+    'build_runner': '^2.5.0',
+    'pubspec_extract': '^2.0.8',
+    'test': '^1.24.0',
+  };
+}